fix reference problems from updating gomap
The material for the universities was in the wrong file.

The callback in RealtimePositionHandler was hooked up in the gomap code
(which is very bad) so that is moved to RealtimePositionHandler.

I also made some structural changes to the Pogo scene in order to
better convey a separation between our code and gomap code.